Moved Enums into GameBoard folder
